// apps/web/src/api/tipos.gerados.ts
export interface components {
    schemas: {
// ...
        CriarExameRequest: {
            /** Categoria */
            categoria: string;
            /** Nome */
            nome: string;
            /** Preco Base */
            preco_base: number | string;
        };
// ...
        CriarRegraPlantaoRequest: {
            /** Dia Semana */
            dia_semana: number;
            /**
             * Hora Inicio
             * Format: time
             */
            hora_inicio: string;
            /**
             * Hora Fim
             * Format: time
             */
            hora_fim: string;
            /** Valor Adicional */
            valor_adicional: number | string;
        };
// ...
        ExameResponse: {
            /** Id */
            id: string;
            /** Categoria */
            categoria: string;
            /** Nome */
            nome: string;
            /** Preco Base */
            preco_base: string;
            /** Ativo */
            ativo: boolean;
        };
// ...
        RegraPlantaoResponse: {
            /** Id */
            id: string;
            /** Dia Semana */
            dia_semana: number;
            /**
             * Hora Inicio
             * Format: time
             */
            hora_inicio: string;
            /**
             * Hora Fim
             * Format: time
             */
            hora_fim: string;
            /** Valor Adicional */
            valor_adicional: string;
            /** Ativo */
            ativo: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}